Add permissions boilerplate + some renames (#96)
