+++ src/hermes/package.json
@@ -25,10 +25,13 @@
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
     "graphql": "^16.6.0",
+    "graphql-middleware": "6.1.33",
+    "graphql-shield": "7.6.4",
     "jsonwebtoken": "^8.5.1",
     "luxon": "^3.1.0",
     "nexus": "^1.3.0",
     "nexus-prisma": "^0.35.0",
+    "lodash": "^4.17.21",
     "prisma": "^4.6.1",
     "ts-node-dev": "^2.0.0"
   },
