+++ src/hermes/src/server/createServer.ts
@@ -1,11 +1,15 @@
 import { ApolloServer } from 'apollo-server-express';
 
-import schema from './schema';
+import { schema } from './schema/makeSchema';
 import { AppContext } from './context';
 import { authenticateUser } from './schema/User/helpers/authentication';
+import { permissions } from './permissions'
+import {applyMiddleware} from 'graphql-middleware';
 // import { authenticateUser } from './schema/User';
 
 type ApolloServerConfig = Omit<AppContext, 'req' | 'res' | 'user'>;
+
+const schemaWithMiddleware = applyMiddleware(schema, permissions);
 
 /**
  *
@@ -17,13 +21,13 @@ type ApolloServerConfig = Omit<AppContext, 'req' | 'res' | 'user'>;
 export const createApolloServer = ({ prisma }: ApolloServerConfig) => {
   // Apollo server object
   return new ApolloServer({
-    schema,
+    schema : schemaWithMiddleware,
     context: ({ req, res }): AppContext => {
       return {
         req,
         res,
         prisma,
-        user: authenticateUser(req.headers.authorization)
+        user: authenticateUser(req.headers.authorization),
       };
     }
   });
