+++ src/hermes/src/server/permissions.ts
@@ -1,0 +1,15 @@
+import { rule, shield } from "graphql-shield"
+import { AppContext } from "./context"
+import { isNil } from 'lodash'
+
+const rules = {
+    isAuthenticatedUser: rule()((_, __, { user } : AppContext) => {
+        return !isNil(user?.token)
+    })
+}
+
+export const permissions = shield(
+    {
+        // Queries with rules go here.
+    }
+)
